// packages/agents-mcp/src/models/notfound.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod";

/**
 * A short code indicating the error code returned.
 */
export const NotFoundCode1$zodSchema = z.enum([
  "not_found",
]).describe("A short code indicating the error code returned.");

export type NotFoundCode1 = z.infer<typeof NotFoundCode1$zodSchema>;

/**
 * A short code indicating the error code returned.
 */
export const NotFoundCode2$zodSchema = z.enum([
  "not_found",
]).describe("A short code indicating the error code returned.");

export type NotFoundCode2 = z.infer<typeof NotFoundCode2$zodSchema>;

/**
 * Legacy error format for backward compatibility.
 */
export type NotFoundError = { code: NotFoundCode2; message: string };

export const NotFoundError$zodSchema: z.ZodType<
  NotFoundError,
  z.ZodTypeDef,
  unknown
> = z.object({
  code: NotFoundCode2$zodSchema,
  message: z.string(),
}).describe("Legacy error format for backward compatibility.");

export type NotFound = {
  title: string;
  status: number;
  detail: string;
  instance?: string | undefined;
  requestId?: string | undefined;
  code: NotFoundCode1;
  error: NotFoundError;
};

export const NotFound$zodSchema: z.ZodType<NotFound, z.ZodTypeDef, unknown> = z
  .object({
    code: NotFoundCode1$zodSchema,
    detail: z.string(),
    error: z.lazy(() => NotFoundError$zodSchema),
    instance: z.string().optional(),
    requestId: z.string().optional(),
    status: z.number().int(),
    title: z.string(),
  });