// packages/agents-mcp/src/models/unauthorized.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod";

/**
 * A short code indicating the error code returned.
 */
export const UnauthorizedCode1$zodSchema = z.enum([
  "unauthorized",
]).describe("A short code indicating the error code returned.");

export type UnauthorizedCode1 = z.infer<typeof UnauthorizedCode1$zodSchema>;

/**
 * A short code indicating the error code returned.
 */
export const UnauthorizedCode2$zodSchema = z.enum([
  "unauthorized",
]).describe("A short code indicating the error code returned.");

export type UnauthorizedCode2 = z.infer<typeof UnauthorizedCode2$zodSchema>;

/**
 * Legacy error format for backward compatibility.
 */
export type UnauthorizedError = { code: UnauthorizedCode2; message: string };

export const UnauthorizedError$zodSchema: z.ZodType<
  UnauthorizedError,
  z.ZodTypeDef,
  unknown
> = z.object({
  code: UnauthorizedCode2$zodSchema,
  message: z.string(),
}).describe("Legacy error format for backward compatibility.");

export type Unauthorized = {
  title: string;
  status: number;
  detail: string;
  instance?: string | undefined;
  requestId?: string | undefined;
  code: UnauthorizedCode1;
  error: UnauthorizedError;
};

export const Unauthorized$zodSchema: z.ZodType<
  Unauthorized,
  z.ZodTypeDef,
  unknown
> = z.object({
  code: UnauthorizedCode1$zodSchema,
  detail: z.string(),
  error: z.lazy(() => UnauthorizedError$zodSchema),
  instance: z.string().optional(),
  requestId: z.string().optional(),
  status: z.number().int(),
  title: z.string(),
});